refactor html and scss
the page width is now consistent everywhere, so it is no longer
necessary for individual pages to be wrapped in a ".narrow-body" div.

also fixed a problem where the page would be able to scroll up and down
20px even when all of the content should fit onto the page
