+++ frontend/src/app/components/app/app.component.scss
@@ -1,0 +1,6 @@
+.boop-header.mat-display-1 {
+    text-align: center;
+    font-size: 42px;
+    padding-top: 20px;
+    margin-bottom: 20px;
+}
