+++ frontend/src/app/components/common/loading-bar/loading-bar.component.html
@@ -1,3 +1,3 @@
-<div class="narrow-body" style="margin-top: 50px;">
+<div style="margin-top: 50px;">
     <mat-progress-bar mode="indeterminate"></mat-progress-bar>
 </div>
