+++ frontend/src/app/components/not-found404/not-found404.component.html
@@ -1,4 +1,4 @@
-<div class="narrow-body" style="text-align: center;">
+<div style="text-align: center;">
     <h1 style="margin-bottom: 15px; font-size: xx-large;">Oh no! 404!</h1>
     <p style="font-size: large; margin-bottom: 15px;">We couldn't find any page at this location.</p>
     <button mat-flat-button color="primary" routerLink="/">Go to Home</button>
