+++ frontend/src/app/components/push-subscribe/push-subscribe.component.html
@@ -1,11 +1,8 @@
-<div class="narrow-body">
-    <h2 style="text-align: center; margin-bottom: 0;">Push Notifications Setup</h2>
-    <p class="menu-paragraph"> Boop uses push notification reminders to remind you and your friends to keep in touch.
-    </p>
-    <button mat-raised-button color="primary" style="margin: auto; display: block; margin-top: 10px;"
-        (click)="activateNotifications()">Activate Notifications On This Device</button>
-    <button mat-button style="margin: auto; display: block;" routerLink="/home">Skip This Step</button>
+<h2 style="text-align: center; margin-bottom: 0;">Push Notifications Setup</h2>
+<p class="menu-paragraph"> Boop uses push notification reminders to remind you and your friends to keep in touch.</p>
+<button mat-raised-button color="primary" style="margin: auto; display: block; margin-top: 10px;"
+    (click)="activateNotifications()">Activate Notifications On This Device</button>
+<button mat-button style="margin: auto; display: block;" routerLink="/home">Skip This Step</button>
 
-    <p style="opacity: 50%; margin-top: 20px; max-width: 2.6in;" class="menu-paragraph">Web push notifications are not
-        currently supported on iOS devices.</p>
-</div>
+<p style="opacity: 50%; margin-top: 20px; max-width: 2.6in;" class="menu-paragraph">Web push notifications are not
+    currently supported on iOS devices.</p>
